requirements.txt
django
django-with-asserts
